+++ Scripts/Audio/Cue.cs
@@ -1,8 +1,10 @@
 using UnityEngine;
 
-namespace Etienne {
+namespace Etienne
+{
     [System.Serializable]
-    public struct Cue {
+    public struct Cue
+    {
         public Sound Sound => new Sound(clips[Random.Range(0, clips.Length)], parameters);
         public AudioClip[] Clips => clips;
         public SoundParameters Parameters => parameters;
@@ -10,19 +12,26 @@ namespace Etienne {
         [SerializeField] private AudioClip[] clips;
         [SerializeField] private SoundParameters parameters;
 
-        public Cue(AudioClip[] clips = null, float volume = 1, float pitch = 1, bool loop = false) {
+        public Cue(AudioClip[] clips = null, float volume = 1, float pitch = 1, bool loop = false)
+        {
             this.clips = clips;
             parameters = new SoundParameters(volume, pitch, loop);
         }
 
-        public Cue(AudioClip[] clips, SoundParameters parameters) {
+        public Cue(AudioClip[] clips, SoundParameters parameters)
+        {
             this.clips = clips;
             this.parameters = parameters;
         }
 
-
-        public static implicit operator Sound(Cue cue) {
+        public static implicit operator Sound(Cue cue)
+        {
             return cue.Sound;
+        }
+
+        public override string ToString()
+        {
+            return clips.Length == 0 ? "Nothing" : $"Cue of {clips.Length} clips";
         }
     }
 }