+++ Scripts/Extensions/Axis.cs
@@ -1,0 +1,1 @@
+﻿public enum Axis { X, Y, Z }