+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Dates in European way
 
+## [0.1.3] - 7/03/2022
+### Added
+- Cinemachine Sample
+- Dotween Sample
+
 ## [0.1.2] - 7/03/2022
 ### Fixed
 - ClearLog Build Bug